[refactoring] Returning a View instead of JSON

+++ src/main/java/com/proteinduo/domain/routineManage/repository/RoutineRepository.java
@@ -1,5 +1,6 @@
 package com.proteinduo.domain.routineManage.repository;
 
+import com.proteinduo.domain.exerciseManage.entity.Exercise;
 import com.proteinduo.domain.memberManage.entity.Member;
 import com.proteinduo.domain.routineManage.entity.Routine;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -20,13 +21,13 @@ import java.util.Optional;
  * 2024-08-11        82102       최초 생성
  */
 @Repository
-public interface RoutineRepository extends JpaRepository<Routine, Long> {
+public interface RoutineRepository extends JpaRepository<Routine, Integer> {
 
     Optional<Routine> findByRoutineId(Integer routineId);
     void deleteByRoutineId(Integer routineId);
     Boolean existsByRoutineId(Integer routineId);
 
     List<Routine> findByMember_MemberId(String memberId);
-
+   // List<Exercise> exercises = routineOptional.get().getExercises();
 }
